+++ ASM/c/triforce.c
@@ -27,7 +27,7 @@ void draw_triforce_count(z64_disp_buf_t* db) {
         frames = frames % (TRIFORCE_SPRITE_FRAMES * FRAMES_PER_CYCLE);
     } else {
         // Do a fade in/out effect if not in pause screen
-        if ( frames <= TRIFORCE_FRAMES_FADE_INTO ) {
+        if (frames <= TRIFORCE_FRAMES_FADE_INTO) {
             alpha = frames * 255 / TRIFORCE_FRAMES_FADE_INTO;
         } else if (frames <= TRIFORCE_FRAMES_FADE_INTO + TRIFORCE_FRAMES_VISIBLE ) {
             alpha = 255;
@@ -65,9 +65,14 @@ void draw_triforce_count(z64_disp_buf_t* db) {
     // Setup draw location
     int str_len = required_digits + pieces_digits + 1;
     int total_w = str_len * font_sprite.tile_w + triforce_sprite.tile_w;
+    // Draw the counter centered horizontally and at the bottom of CRT safe screen space.
     int draw_x = Z64_SCREEN_WIDTH / 2 - total_w / 2;
-    int draw_y_text = Z64_SCREEN_HEIGHT - (font_sprite.tile_h * 1.5) + 1;
-    int draw_y_triforce = Z64_SCREEN_HEIGHT - (triforce_sprite.tile_h * 1.5) + 3 + 1;
+    // And on pause screen, keep the height but move it on the right, symmetric to the rupee icon.
+    if (z64_game.pause_ctxt.state > PAUSE_STATE_OFF) {
+        draw_x = (Z64_SCREEN_WIDTH - 26) - total_w;
+    }
+    int draw_y_text = 206;
+    int draw_y_triforce = 206;
 
     // Create collected/required string
     char text[str_len + 1];