+++ Messages.py
@@ -360,7 +360,7 @@ ITEM_MESSAGES: list[tuple[int, str]] = [
     (0x9342, "\x08You found the \x05\x41Zora Region Enemy Souls\x05\x40!"),
     (0x9343, "\x08You found the \x05\x41Lon Lon Ranch Enemy Souls\x05\x40!"),
     (0x9344, "\x08You found the \x05\x41Grotto Enemy Souls\x05\x40!"),
-    
+
     (0x9500, "\x08You found a \x05\x41Fishing Rod\x05\x40!"),
     (0x9501, "\x08You found a \x05\x41Fish (Child 6lb)\x05\x40!"),
     (0x9502, "\x08You found a \x05\x41Fish (Child 8lb)\x05\x40!"),
@@ -1238,7 +1238,7 @@ def add_item_messages(messages: list[Message], shop_items: Iterable[ShopItem], w
 
 
 def find_message_index(messages: list[Message], id):
-    
+
     for i in range(0, len(messages)):
         if messages[i].id == id:
             return i
@@ -1460,20 +1460,19 @@ def update_warp_song_text(messages: list[Message], world: World) -> None:
         0x4004: 'LH Owl Flight -> Hyrule Field',
     }
 
-    if world.settings.logic_rules != "glitched": # Entrances not set on glitched logic so following code will error
-        for id, entr in msg_list.items():
-            if 'warp_songs_and_owls' in world.settings.misc_hints or not world.settings.warp_songs:
-                destination = world.get_entrance(entr).connected_region
-                destination_name = HintArea.at(destination)
-                color = COLOR_MAP[destination_name.color]
-                if destination_name.preposition(True) is not None:
-                    destination_name = f'to {destination_name}'
-            else:
-                destination_name = 'to a mysterious place'
-                color = COLOR_MAP['White']
+    for id, entr in msg_list.items():
+        if 'warp_songs_and_owls' in world.settings.misc_hints or not world.settings.warp_songs:
+            destination = world.get_entrance(entr).connected_region
+            destination_name = HintArea.at(destination)
+            color = COLOR_MAP[destination_name.color]
+            if destination_name.preposition(True) is not None:
+                destination_name = f'to {destination_name}'
+        else:
+            destination_name = 'to a mysterious place'
+            color = COLOR_MAP['White']
 
-            new_msg = f"\x08\x05{color}Warp {destination_name}?\x05\40\x09\x01\x01\x1b\x05\x42OK\x01No\x05\40"
-            update_message_by_id(messages, id, new_msg)
+        new_msg = f"\x08\x05{color}Warp {destination_name}?\x05\40\x09\x01\x01\x1b\x05\x42OK\x01No\x05\40"
+        update_message_by_id(messages, id, new_msg)
 
     if world.settings.owl_drops:
         for id, entr in owl_messages.items():